auto dir creation and copying of resources to designated folder
Since local images need relative path we now copy the local images to a resource folder. for that when saving a file we create a folder called temp_output. in there we create the resourcce folder and copy the image into it. after the (html)-files are written etc we rename the temp_output to PROJECTNAME_output